+++ src/CustomizeBuildings/Properties/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("CustomizeBuildings")]
 [assembly: AssemblyProduct("CustomizeBuildings")]
-[assembly: AssemblyVersion("1.0.123.0")]
-[assembly: AssemblyFileVersion("1.0.123.0")]
+[assembly: AssemblyVersion("1.0.124.0")]
+[assembly: AssemblyFileVersion("1.0.124.0")]